backend/requirement.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.9.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.2
httpx==0.27.2
email-validator==2.2.0
